+++ frontend/garticband/src/lib/components/SelectInstrument.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-    import Drums from '$lib/instruments/Drums.svelte'
-    import Bass from '$lib/instruments/Bass.svelte'
-    import Piano from '$lib/instruments/Piano.svelte'
-    import Synth from '$lib/instruments/Synth.svelte'
+    import Drums from '$lib/components/instruments/Drums.svelte'
+    import Bass from '$lib/components/instruments/Bass.svelte'
+    import Piano from '$lib/components/instruments/Piano.svelte'
+    import Synth from '$lib/components/instruments/Synth.svelte'
 
-    import { type SelectInstrumentProps } from './types'
+    import { type SelectInstrumentProps } from '../types'
 
     const { instrumentId, ...others }: SelectInstrumentProps = $props()
 </script>
