+++ frontend/garticband/src/lib/components/Sequencer.svelte
