+++ frontend/garticband/src/lib/components/UpNext.svelte
