+++ frontend/garticband/src/lib/components/instruments/Bass.svelte
