+++ frontend/garticband/src/lib/components/instruments/Piano.svelte
