+++ frontend/garticband/src/lib/components/instruments/Synth.svelte
