+++ frontend/garticband/src/routes/game/dummyComposition/+page.svelte
@@ -1,6 +1,6 @@
 <script>
     import { allFalseRoomState } from "$lib/dummyData";
     import { randomRoomState } from "$lib/dummyData";
-    import FinalComposition from "$lib/FinalComposition.svelte";
+    import FinalComposition from "$lib/components/FinalComposition.svelte";
 </script>
 <FinalComposition roomState={allFalseRoomState} />
